Fix incorrect volume averages (change float to double). JSON-ish output in ascii output from IndexFile.

+++ include/bd/volume/fileblock.h
@@ -27,13 +27,13 @@ namespace bd
 ///   block dims Y      | 8 bytes unsigned
 ///   block dims Z      | 8 bytes unsigned
 ///   --
-///   block X pos       | 4 bytes float
-///   block Y pos       | 4 bytes float
-///   block Z pos       | 4 bytes float
+///   block X pos       | 8 bytes float
+///   block Y pos       | 8 bytes float
+///   block Z pos       | 8 bytes float
 ///   --
-///   max val           | 4 bytes float
-///   min val           | 4 bytes float
-///   avg val           | 4 bytes float
+///   max val           | 8 bytes float
+///   min val           | 8 bytes float
+///   avg val           | 8 bytes float
 ///
 ///   isEmpty           | 4 bytes unsigned
 ///   -----------------------------------------
@@ -58,10 +58,10 @@ namespace bd
     uint64_t block_index;    ///< The 1D idx of this block (derived from the i,j,k block-grid coordinates).
     uint64_t data_offset;    ///< Offset into the raw file that the block data starts.
     uint64_t voxel_dims[3];  ///< Dimensions of this block in voxels.
-    float world_pos[3];      ///< Cordinates within canonical cube.
-    float min_val;           ///< The min value found in this block.
-    float max_val;           ///< The largest value found in this block.
-    float avg_val;           ///< Average value within this block.
+    double world_pos[3];      ///< Cordinates within canonical cube.
+    double min_val;           ///< The min value found in this block.
+    double max_val;           ///< The largest value found in this block.
+    double avg_val;           ///< Average value within this block.
     uint32_t is_empty;        ///< If this block is empty or not.
 
   };
